Changed the library to produce a UINT16 so that one decimal point can stilll be obtained without using float.

+++ NonBTinyDHT.h
@@ -76,7 +76,7 @@ public:
    * @brief Reads the humidity from the device
    * @return Returns the humidity read from the device
    */
-  uint8_t readHumidity(void);
+  uint16_t readHumidity(void);
 
   void update(void);
